Rename prioritymechanisms to mechanismpriority
